+++ app/Opulence/Applications/Application.php
@@ -13,7 +13,6 @@ use Exception;
 use Opulence\Applications\Environments\Environment;
 use Opulence\Applications\Tasks\Dispatchers\IDispatcher;
 use Opulence\Applications\Tasks\TaskTypes;
-use Opulence\Ioc\IContainer;
 use RuntimeException;
 
 /**
@@ -23,34 +22,21 @@ class Application
 {
     /** @var string The current Opulence version */
     private static $version = "1.0.0-alpha9";
-    /** @var Paths The paths to various directories used by Opulence */
-    private $paths = null;
     /** @var IDispatcher The task dispatcher */
     private $taskDispatcher = null;
     /** @var Environment The environment the application is running on */
     private $environment = null;
-    /** @var IContainer The IoC container */
-    private $container = null;
     /** @var bool Whether or not the application is currently running */
     private $isRunning = false;
 
     /**
-     * @param Paths $paths The paths to various directories used by Opulence
      * @param IDispatcher $taskDispatcher The task dispatcher
      * @param Environment $environment The current environment
-     * @param IContainer $container The IoC container
      */
-    public function __construct(
-        Paths $paths,
-        IDispatcher $taskDispatcher,
-        Environment $environment,
-        IContainer $container
-    ) {
-        // Order here is important
-        $this->setPaths($paths);
+    public function __construct(IDispatcher $taskDispatcher, Environment $environment)
+    {
         $this->taskDispatcher = $taskDispatcher;
         $this->setEnvironment($environment);
-        $this->container = $container;
     }
 
     /**
@@ -70,22 +56,6 @@ class Application
     }
 
     /**
-     * @return IContainer
-     */
-    public function getIocContainer()
-    {
-        return $this->container;
-    }
-
-    /**
-     * @return Paths
-     */
-    public function getPaths()
-    {
-        return $this->paths;
-    }
-
-    /**
      * @return bool
      */
     public function isRunning()
@@ -99,14 +69,6 @@ class Application
     public function setEnvironment(Environment $environment)
     {
         $this->environment = $environment;
-    }
-
-    /**
-     * @param Paths $paths
-     */
-    public function setPaths($paths)
-    {
-        $this->paths = $paths;
     }
 
     /**